Fix FPU corruption on task switch
There was a clear issue in `fpu_switch`, and more to the initialization
of the FPU than I assumed.
This fixes time not passing on VirtualBox and real hardware, which made
it look as if SnowflakeOS had crashed there, when it was simply frozen.
The bug that remains, on VirtualBox at least, is somewhere in the PS/2
controller: the mouse works until a key is pressed, after which the
mouse breaks too.

+++ kernel/include/kernel/timer.h
@@ -5,7 +5,7 @@
 void init_timer();
 void timer_callback();
 uint32_t timer_get_tick();
-double timer_get_time();
+float timer_get_time();
 void timer_register_callback(handler_t handler);
 void timer_remove_callback(handler_t handler);
